Isolate LoggingImagingListener into a seperate class
